# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "MyApp",
        "version": "1.0.0"
    },
    "paths": {
        "/users": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create User",
                "operationId": "create_user_users_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseSchema_UserRead_"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List Users",
                "operationId": "list_users_users_get",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "description": "页码，从1开始",
                            "default": 1,
                            "title": "Page"
                        },
                        "description": "页码，从1开始"
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "description": "每页条数",
                            "default": 10,
                            "title": "Page Size"
                        },
                        "description": "每页条数"
                    },
                    {
                        "name": "keyword",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "搜索关键词",
                            "title": "Keyword"
                        },
                        "description": "搜索关键词"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseSchema_PageResult_UserRead__"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Me",
                "description": "获取当前登录用户信息（需要 token）",
                "operationId": "get_me_users_me_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseSchema_UserRead_"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/users/{user_id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get User",
                "operationId": "get_user_users__user_id__get",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseSchema_UserRead_"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{user_id}/roles": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "给用户分配角色",
                "operationId": "assign_roles_to_user_users__user_id__roles_put",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "type": "integer"
                                },
                                "title": "Role Ids"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseSchema_UserRead_"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "查看用户的角色列表",
                "operationId": "get_user_roles_users__user_id__roles_get",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "User Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseSchema_list_UserWithRolesRead__"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PageResult_UserRead_": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/UserRead"
                        },
                        "type": "array",
                        "title": "Items",
                        "default": []
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total",
                        "default": 0
                    },
                    "page": {
                        "type": "integer",
                        "title": "Page",
                        "default": 1
                    },
                    "page_size": {
                        "type": "integer",
                        "title": "Page Size",
                        "default": 20
                    }
                },
                "type": "object",
                "title": "PageResult[UserRead]"
            },
...
            "PermissionRead": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "code": {
                        "type": "string",
                        "title": "Code"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "code",
                    "name",
                    "description"
                ],
                "title": "PermissionRead"
            },
...
            "ResponseSchema_PageResult_UserRead__": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "success"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/PageResult_UserRead_"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "title": "ResponseSchema[PageResult[UserRead]]"
            },
...
            "ResponseSchema_UserRead_": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "success"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/UserRead"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "title": "ResponseSchema[UserRead]"
            },
...
            "ResponseSchema_list_UserWithRolesRead__": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "title": "Code",
                        "default": 200
                    },
                    "message": {
                        "type": "string",
                        "title": "Message",
                        "default": "success"
                    },
                    "data": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/UserWithRolesRead"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Data"
                    }
                },
                "type": "object",
                "title": "ResponseSchema[list[UserWithRolesRead]]"
            },
...
            "RoleRead": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "code": {
                        "type": "string",
                        "title": "Code"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Description"
                    },
                    "permissions": {
                        "items": {
                            "$ref": "#/components/schemas/PermissionRead"
                        },
                        "type": "array",
                        "title": "Permissions",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "code",
                    "name",
                    "description"
                ],
                "title": "RoleRead"
            },
...
            "UserCreate": {
                "properties": {
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "email",
                    "password"
                ],
                "title": "UserCreate"
            },
            "UserRead": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "username",
                    "email",
                    "is_active"
                ],
                "title": "UserRead"
            },
            "UserWithRolesRead": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "roles": {
                        "items": {
                            "$ref": "#/components/schemas/RoleRead"
                        },
                        "type": "array",
                        "title": "Roles",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "username",
                    "email",
                    "is_active"
                ],
                "title": "UserWithRolesRead"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "ctx": {
                        "type": "object",
                        "title": "Context"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/api/v1/auth/login"
                    }
                }
            }
        }
    }
}